Added error handling for whenever deleted resource is used in different table

+++ src/main/java/com/project/crud/services/UsersService.java
@@ -2,8 +2,10 @@ package com.project.crud.services;
 
 import com.project.crud.dtos.UsersDto;
 import com.project.crud.entities.Users;
+import com.project.crud.exceptionHandling.ForeignKeyFoundException;
 import com.project.crud.exceptionHandling.ResourceNotFoundException;
 import com.project.crud.mappers.UsersMapper;
+import com.project.crud.repositories.ReviewsRepository;
 import com.project.crud.repositories.UsersRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,10 +19,12 @@ import java.util.stream.StreamSupport;
 public class UsersService {
     private final UsersRepository usersRepository;
     private final UsersMapper usersMapper;
+    private final ReviewsRepository reviewsRepository;
 
-    public UsersService(UsersRepository usersRepository, UsersMapper usersMapper) {
+    public UsersService(UsersRepository usersRepository, UsersMapper usersMapper, ReviewsRepository reviewsRepository) {
         this.usersRepository = usersRepository;
         this.usersMapper = usersMapper;
+        this.reviewsRepository = reviewsRepository;
     }
 
     public List<UsersDto> getAllUsers(){
@@ -29,12 +33,16 @@ public class UsersService {
     }
 
     public UsersDto getUser(String username){
-        return usersMapper.toDto(usersRepository.findById(username).orElseThrow(() -> new ResourceNotFoundException("user", username)));
+        return usersMapper.toDto(usersRepository.findById(username).orElseThrow(
+                () -> new ResourceNotFoundException("user", username)));
     }
 
     public HttpStatus deleteUser(String username){
         if (usersRepository.findById(username).isEmpty()){
             throw new ResourceNotFoundException("user", username);
+        }
+        if (!reviewsRepository.findByIdUsername(username).isEmpty()) {
+            throw new ForeignKeyFoundException("user", username, "Reviews");
         }
         usersRepository.deleteById(username);
         return HttpStatus.OK;
